+++ packages/delir-core/src/engine/IRenderingStreamObserver.ts
