+++ packages/delir-core/src/engine/Task/ClipRenderTask.ts
@@ -1,17 +1,18 @@
-import { AssetPointerScheme } from '@ragg/delir-core/src/project/scheme/keyframe'
 import * as _ from 'lodash'
-import * as KeyframeHelper from '../../../helper/keyframe-helper'
-import { TypeDescriptor } from '../../../plugin-support/type-descriptor'
-import { Clip } from '../../../project'
-import { Expression } from '../../../values'
-import * as RendererFactory from '../../renderer'
-import { IRenderer } from '../../renderer/renderer-base'
+
+import * as KeyframeHelper from '../../helper/keyframe-helper'
+import { TypeDescriptor } from '../../plugin-support/type-descriptor'
+import { Clip } from '../../project'
+import { AssetPointerScheme } from '../../project/scheme/keyframe'
+import { Expression } from '../../values'
+import { RealParameterValues, RealParameterValueTypes } from '../Engine'
 import { compileTypeScript } from '../ExpressionCompiler'
-import { ContextSource } from '../ExpressionContext'
 import * as ExpressionContext from '../ExpressionContext'
+import { ContextSource } from '../ExpressionContext'
 import ExpressionVM from '../ExpressionVM'
-import { RealParameterValues, RealParameterValueTypes } from '../pipeline'
-import RenderRequest from '../render-request'
+import * as RendererFactory from '../renderer'
+import { IRenderer } from '../renderer/renderer-base'
+import RenderRequest from '../RenderRequest'
 import EffectRenderTask from './EffectRenderTask'
 
 export default class ClipRenderTask {
