+++ packages/delir-core/src/engine/renderer/Adjustment/Adjustment.ts
@@ -1,7 +1,7 @@
 import Type from '../../../plugin-support/type-descriptor'
 import { TypeDescriptor } from '../../../plugin-support/type-descriptor'
-import PreRenderingRequest from '../../pipeline/pre-rendering-request'
-import RenderingRequest from '../../pipeline/render-request'
+import PreRenderingRequest from '../../PreRenderingRequest'
+import RenderingRequest from '../../RenderRequest'
 import { IRenderer } from '../renderer-base'
 
 import Asset from '../../../project/asset'
