+++ packages/delir-core/src/engine/renderer/Audio/Audio.ts
@@ -3,8 +3,8 @@ import * as _ from 'lodash'
 import { resampling } from '../../../helper/Audio'
 import Type from '../../../plugin-support/type-descriptor'
 import { TypeDescriptor } from '../../../plugin-support/type-descriptor'
-import PreRenderingRequest from '../../pipeline/pre-rendering-request'
-import RenderingRequest from '../../pipeline/render-request'
+import PreRenderingRequest from '../../PreRenderingRequest'
+import RenderingRequest from '../../RenderRequest'
 import { IRenderer } from '../renderer-base'
 
 import Asset from '../../../project/asset'
