+++ packages/delir-core/src/engine/renderer/Image/Image.ts
@@ -1,8 +1,8 @@
 import * as _ from 'lodash'
 import Type from '../../../plugin-support/type-descriptor'
 import { TypeDescriptor } from '../../../plugin-support/type-descriptor'
-import PreRenderingRequest from '../../pipeline/pre-rendering-request'
-import RenderingRequest from '../../pipeline/render-request'
+import PreRenderingRequest from '../../PreRenderingRequest'
+import RenderingRequest from '../../RenderRequest'
 import { IRenderer } from '../renderer-base'
 
 import Asset from '../../../project/asset'
