+++ packages/delir-core/src/index.ts
@@ -1,11 +1,11 @@
-import * as Engine from './engine/index'
+import * as Engine from './engine'
 import * as Exceptions from './exceptions'
 import * as PluginSupport from './plugin-support'
 import * as Project from './project/index'
 import * as Values from './values'
 
-import PreRenderRequest from './engine/pipeline/pre-rendering-request'
-import RenderRequest from './engine/pipeline/render-request'
+import PreRenderRequest from './engine/PreRenderingRequest'
+import RenderRequest from './engine/RenderRequest'
 import PluginBase from './plugin-support/plugin-base'
 import PluginRegistry from './plugin-support/plugin-registry'
 import PostEffectBase from './plugin-support/PostEffectBase'
